refactor(server): move edit_recipe and open_clip_compat to utils/
