+++ server/src/service_metadata.py
@@ -16,7 +16,7 @@ from llm_provider_ollama import OllamaProvider
 from llm_provider_lmstudio import LMStudioProvider
 from llm_provider_chatgpt import ChatGPTProvider
 from llm_provider_gemini import GeminiProvider
-from edit_recipe import filter_edit_recipe_by_controls
+from utils.edit_recipe import filter_edit_recipe_by_controls
 from config import logger, DEFAULT_METADATA_PROVIDER, DEFAULT_METADATA_LANGUAGE
 import service_training as training_service
 from PIL import Image
