+++ server/src/utils/open_clip_compat.py
